increasing precision

+++ script/strategies/kandel/KandelDeployer.s.sol
@@ -46,8 +46,8 @@ contract KandelDeployer is Deployer {
     uint compoundRateBase,
     uint compoundRateQuote
   ) public {
-    require(uint16(compoundRateBase) == compoundRateBase, "compoundRateBase is too big");
-    require(uint16(compoundRateQuote) == compoundRateQuote, "compoundRateQuote is too big");
+    require(uint24(compoundRateBase) == compoundRateBase, "compoundRateBase is too big");
+    require(uint24(compoundRateQuote) == compoundRateQuote, "compoundRateQuote is too big");
     IMangrove mgv = IMangrove(fork.get("Mangrove"));
     (MgvStructs.GlobalPacked global,) = mgv.config(address(0), address(0));
 
@@ -62,7 +62,7 @@ contract KandelDeployer is Deployer {
     );
 
     broadcast();
-    current.setCompoundRates(uint16(compoundRateBase), uint16(compoundRateQuote));
+    current.setCompoundRates(uint24(compoundRateBase), uint24(compoundRateQuote));
 
     string memory kandelName = getName(IERC20(base), IERC20(quote));
     fork.set(kandelName, address(current));
